fix issue #469 (#471)
