+++ zrpc/internal/rpcpubserver_test.go
@@ -1,0 +1,33 @@
+package internal
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/tal-tech/go-zero/core/netx"
+)
+
+func TestFigureOutListenOn(t *testing.T) {
+	tests := []struct {
+		input  string
+		expect string
+	}{
+		{
+			input:  "192.168.0.5:1234",
+			expect: "192.168.0.5:1234",
+		},
+		{
+			input:  "0.0.0.0:8080",
+			expect: netx.InternalIp() + ":8080",
+		},
+		{
+			input:  ":8080",
+			expect: netx.InternalIp() + ":8080",
+		},
+	}
+
+	for _, test := range tests {
+		val := figureOutListenOn(test.input)
+		assert.Equal(t, test.expect, val)
+	}
+}
